refactor : Improve Card Component modularity
